Auto-Metadata-to-Cosmos-Task3/requirements.txt
azure-core==1.36.0
azure-cosmos==4.14.2
azure-functions==1.24.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
MarkupSafe==3.0.3
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
